Split ContractHelper to allow custom debugging experience in CoreRT.
